ref(kb): relocate LanceDB cleanup filters to the storage layer
